+++ packages/stream-model-instance-document-handler/src/__tests__/model-instance-document.test.ts
