+++ packages/stream-model-instance-document-handler/src/model-instance-document-handler.ts
